Move LogHelper into EvaluateLogger.
